+++ schema/site_stringdata.go
@@ -562,7 +562,14 @@ const SiteSchemaJSON = `{
     "auth.enableUsernameChanges": {
       "description": "Enables users to change their username after account creation. Warning: setting this to be true has security implications if you have enabled (or will at any point in the future enable) repository permissions with an option that relies on username equivalency between Sourcegraph and an external service or authentication provider. Do NOT set this to true if you are using non-built-in authentication OR rely on username equivalency for repository permissions.",
       "type": "boolean",
-      "default": false
+      "default": false,
+      "group": "Authentication"
+    },
+    "auth.minPasswordLength": {
+      "description": "The minimum number of Unicode code points that a password must contain.",
+      "type": "integer",
+      "default": 12,
+      "group": "Authentication"
     },
     "update.channel": {
       "description": "The channel on which to automatically check for Sourcegraph updates.",